move and rename ThymeTable to TimeTable
