+++ src/sections/TimeSheet/components/Table/index.js
@@ -11,14 +11,14 @@ import { getDurationRounding, getDurationAmount, getRoundingOn } from 'sections/
 
 import { addProject } from 'sections/Projects/actions';
 
-import { updateTime, removeTime } from '../../actions/time';
+import { updateTime, removeTime } from '../../../../actions/time';
 
-import { getDateSort } from '../../selectors/time';
+import { getDateSort } from '../../../../selectors/time';
 
-import NewTime from './New';
-import Entry from './Entry';
+import NewTime from '../Entry/New';
+import Entry from '../Entry/Entry';
 
-type ThymeTableType = {
+type TimeTableType = {
   sort: sortDirection;
   entries: Array<timeType>;
   now: Date;
@@ -29,7 +29,7 @@ type ThymeTableType = {
   onAddProject: (project: string) => string;
 };
 
-function ThymeTable({
+function TimeTable({
   sort,
   entries,
   now,
@@ -38,7 +38,7 @@ function ThymeTable({
   onEntryUpdate,
   onEntryRemove,
   onAddProject,
-}: ThymeTableType) {
+}: TimeTableType) {
   const New = (
     <NewTime
       now={now}
@@ -135,4 +135,4 @@ function mapDispatchToProps(dispatch: Dispatch<*>) {
   };
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(ThymeTable);
+export default connect(mapStateToProps, mapDispatchToProps)(TimeTable);
