+++ src/sections/TimeSheet/index.js
@@ -10,9 +10,9 @@ import Accordion from 'semantic-ui-react/dist/commonjs/modules/Accordion/Accordi
 import Responsive from 'semantic-ui-react/dist/commonjs/addons/Responsive/Responsive';
 import Pagination from 'semantic-ui-react/dist/commonjs/addons/Pagination/Pagination';
 
-import DateRange from './components/DateRange/index';
-import DateSort from './components/DateSort/index';
-import ThymeTable from '../../components/ThymeTable/index';
+import DateRange from './components/DateRange';
+import DateSort from './components/DateSort';
+import TimeTable from './components/Table';
 
 import { changePage } from '../../actions/time';
 
@@ -82,7 +82,7 @@ class TimeSheet extends Component<TimeSheetProps, TimeSheetState> {
             <DateSort />
           </div>
         </Responsive>
-        <ThymeTable
+        <TimeTable
           entries={entries.filter((item, index) => index <= end && index >= start)}
           now={now}
         />
